refactor: extract ElementNode.css.ts from overlays
